client and server are both compiling again. Merged changes

+++ protocol/common/src/main/java/edu/drexel/group5/PacketFactory.java
@@ -148,12 +148,12 @@ public class PacketFactory {
 		return new DatagramPacket(data, data.length, destination);
 	}
 
-	public DatagramPacket createStreamMessage(byte sessionId, byte sequenceNumber, byte[] data, byte[] crc) throws SocketException, IOException {
+	public DatagramPacket createStreamMessage(byte sessionId, int sequenceNumber, byte[] data, byte[] crc) throws SocketException, IOException {
 		final ByteArrayOutputStream bytesOut = new ByteArrayOutputStream();
 		final DataOutputStream outputStream = new DataOutputStream(bytesOut);
 		outputStream.writeByte(MessageType.STREAM.getMessageId());
 		outputStream.writeByte(sessionId);
-		outputStream.writeByte(sequenceNumber);
+		outputStream.writeInt(sequenceNumber);
 		outputStream.writeInt(data.length);
 		outputStream.write(data, 0, data.length);
 		outputStream.writeInt(crc.length);
